Change indent size for XML files
We switch to 2-size space indents instead of 4-size for XML files. The reasons:

a. The same rules are used for major projects:
	https://github.com/dotnet/aspnetcore/blob/0cde903fa8a45bbfe892a3afae7650f29edd6d24/.editorconfig#L58
	https://github.com/dotnet/wpf/blob/8a374027abd1f7cb0561d040c37a87ff1bf24ce8/.editorconfig#L146
b. I don't find recommendations for using 4 spaces indents for XML files. It's either tabs or 2 spaces indents.

We keep using 4-size space indents for XAML files. This is a style used by major WPF projects:

https://github.com/dotnet/wpf/tree/main/src/Microsoft.DotNet.Wpf/src/Themes/XAML
https://github.com/MaterialDesignInXAML/MaterialDesignInXamlToolkit/tree/master/MaterialDesignThemes.Wpf/Themes
https://github.com/telerik/xaml-sdk